Make it actually find matching iTerm2 windows
Matching on name begins and ends with doesn't work, I'm not sure why,
but contains works and seems no worse than the grep pattern for linux

Unfortunately, using contains always matches the current window, so
we need to avoid returning the current window as the active window,
hence the hacky code to get the last match as the active window.

I really, really don't like the code for this, but it works. Another
option might be to use the value of titleold in the find_win_cmd to
exclude the current window, but I'm not sure how reliable it would be.